Modularize views folder
