+++ backend/ilm/views/__init__.py
@@ -1,8 +1,6 @@
-from .views import (
-    CourseViewSet,
-    ModuleViewSet,
-    ContentViewSet
-)
+from .view_courses import CourseViewSet
+from .view_modules import ModuleViewSet
+from .view_content import ContentViewSet
 
 
 __all__ = ['CourseViewSet', 'ModuleViewSet', 'ContentViewSet']
